fix(favorites): fixed focus event
There are scss hover styles around the favorite heart for displaying the proper colors. However, when you add a button or role of button to an html element it will still apply the hover styles until the item is unfocused. To allievate this effect, we can simply just add a mouse over handler that prevents the focus event from happening when clicked. a11y still works as when the component is higlighted or focused it will work from when the user moves the current tabIndex to the component