refactor(spaces): allows space components to be called without a space id
In some cases we just want a single space or payers information. In this event, we don't want to have to prop drill down the spaceId or payerid to all the components that are usign spaces components. Rather we can just pass it undefined or nothing and it will just fetch the first space in the spaces array. It will throw a console warning if we are attempting to use this when there is more than 1 space or payer in the array. Not recommended